To try to work around [#85026532], implementing a hack workaround because you can't `-setNilValueForKey:` on an NS(Mutable)Dictionary.
The QuickDialog stuff allows the data object to be pretty much anything that is KVC-compliant, but it's not being smart enough about nil. It seems this is a reported problem: https://github.com/escoz/QuickDialog/issues/392  but I see zero movement in the almost 2 years since the problem was reported. The workaround is tough to do here because our data object is an `NSMutableDictionary` and well... that's tough to want to extend and subclas in this way.

So... we'll just guard against the data object being a dictionary, and just not do it in that case -- it's functionally the same end result anyways.